frontend/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.40.1
azure-storage-blob==12.24.0
